+++ lib/src/deluxe_bottom_bar.dart
@@ -105,9 +105,7 @@ class _DeluxeBottomBarState extends State<DeluxeBottomBar>
           _ActionButton(
             onActionButtonTapped: widget.onActionButtonTapped,
             backgroundColor: widget.actionButtonStyle.backgroundColor,
-            iconColor: widget.actionButtonStyle.iconColor,
-            icon: widget.actionButtonStyle.icon,
-            iconSize: widget.buttonStyle.iconSize,
+            widget: widget.actionButtonStyle.widget,
             boxShadow: widget.style.boxShadow,
           ),
         ],